feat: add category management and inventory module
- Register new inventory-related activities in Manifest.
- Implement CategoriaCarpetaAdapter and CategoriasActivity.
- Add inventory item support for weighted products in DB.
- Update navigation to use InventarioMainActivity.

+++ app/src/main/java/com/aplicaion/minimarketapp/api/MinimarketApiClient.kt
@@ -187,7 +187,7 @@ class MinimarketApiClient(private val context: Context) {
                     )
                 )
                 val currentProd = productoDao.getByIdSync(item.producto.id)
-                val nuevoStock = ((currentProd?.stock ?: item.producto.stock) - item.cantidad).coerceAtLeast(0)
+                val nuevoStock = ((currentProd?.stock ?: item.producto.stock) - item.cantidad).toInt().coerceAtLeast(0)
                 productoDao.updateStock(item.producto.id, nuevoStock)
             }
